Refactor ingestToFeature api request

+++ io.datasearch.epidatafuse.core/src/main/java/io/datasearch/epidatafuse/core/RequestHandler.java
@@ -135,12 +135,16 @@ public class RequestHandler {
 
         Path rootDir = Paths.get("public", "uploads", pipelineName, featureName);
         try {
-            Files.createDirectories(rootDir);
-            Files.copy(file.getInputStream(), rootDir.resolve(file.getOriginalFilename()));
-            File zipfile = new File(rootDir.resolve(file.getOriginalFilename()).toString());
-            ZipFile zipFile = new ZipFile(zipfile);
-            Files.createDirectories(rootDir.resolve("shapefile"));
-            zipFile.extractAll(rootDir.resolve("shapefile").toString());
+            if (!Files.exists(rootDir.resolve(file.getOriginalFilename()))) {
+                Files.createDirectories(rootDir);
+                Files.copy(file.getInputStream(), rootDir.resolve(file.getOriginalFilename()));
+                if ("application/zip".equals(file.getContentType())) {
+                    File zipfile = new File(rootDir.resolve(file.getOriginalFilename()).toString());
+                    ZipFile zipFile = new ZipFile(zipfile);
+                    Files.createDirectories(rootDir.resolve("shapefile"));
+                    zipFile.extractAll(rootDir.resolve("shapefile").toString());
+                }
+            }
         } catch (Exception e) {
             throw new RuntimeException("Could not store the file. Error: " + e.getMessage());
         }
@@ -202,8 +206,9 @@ public class RequestHandler {
     public String ingestToFeature(@RequestBody Map<String, Object> payload) {
         Response response;
         try {
-            if (payload.get(PipelineUtil.PIPELINE_NAME_KEY) != null) {
-                String pipelineName = (String) payload.get(PipelineUtil.PIPELINE_NAME_KEY);
+            String pipelineName = (String) payload.get(PIPELINE_NAME_KEY);
+            if (pipelineName != null
+                    && ServerContext.getPipeline(pipelineName) != null) {
                 IngestConfig ingestConfig = new IngestConfig(payload);
                 Boolean status = FusionPipeLineController.ingestToFeature(pipelineName, ingestConfig);
                 if (status) {
@@ -439,7 +444,6 @@ public class RequestHandler {
             }
         }
     }
-
 
     @RequestMapping("/getConversionMethodInfo")
     public String getConversionMethodInfo(@RequestBody Map<String, Object> payload) {
